feat(waiting): 소켓 핸들러 분리 및 의존성 주입 적용
